# docs/swagger.yaml
definitions:
...
  handler.action:
    properties:
      id:
        example: edit
        type: string
    required:
    - id
    type: object
...
  handler.authByActionsRequest:
    properties:
      actions:
        items:
          $ref: '#/definitions/handler.action'
        type: array
      resources:
        description: can't be empty
        items:
          $ref: '#/definitions/handler.resource'
        type: array
      subject:
        $ref: '#/definitions/handler.subject'
        type: object
      system:
        example: bk_paas
        type: string
    required:
    - actions
    - resources
    - subject
    - system
    type: object
  handler.authByActionsResponse:
    additionalProperties:
      type: boolean
    type: object
  handler.authByResourcesRequest:
    properties:
      action:
        $ref: '#/definitions/handler.action'
        type: object
      resources_list:
        items:
          items:
            $ref: '#/definitions/handler.resource'
          type: array
        type: array
      subject:
        $ref: '#/definitions/handler.subject'
        type: object
      system:
        example: bk_paas
        type: string
    required:
    - action
    - subject
    - system
    type: object
  handler.authByResourcesResponse:
    additionalProperties:
      type: boolean
    type: object
  handler.authRequest:
    properties:
      action:
        $ref: '#/definitions/handler.action'
        type: object
      resources:
        description: required
        items:
          $ref: '#/definitions/handler.resource'
        type: array
      subject:
        $ref: '#/definitions/handler.subject'
        type: object
      system:
        example: bk_paas
        type: string
    required:
    - action
    - resources
    - subject
    - system
    type: object
  handler.authResponse:
    properties:
      allowed:
        example: false
        type: boolean
    type: object
...
  handler.policyIDSubject:
    properties:
      id:
        example: 100
        type: integer
      subject:
        $ref: '#/definitions/handler.policyResponseSubject'
        type: object
    type: object
  handler.policyListResponse:
    properties:
      count:
        example: 120
        type: integer
      metadata:
        $ref: '#/definitions/handler.policyListResponseMetadata'
        type: object
      results:
        items:
          $ref: '#/definitions/handler.thinPolicyResponse'
        type: array
    type: object
  handler.policyListResponseMetadata:
    properties:
      action:
        $ref: '#/definitions/handler.policyResponseAction'
        type: object
      system:
        example: bk_test
        type: string
      timestamp:
        example: 1592899208
        type: integer
    type: object
  handler.policyResponseAction:
    properties:
      id:
        example: edit
        type: string
    type: object
  handler.policyResponseSubject:
    properties:
      id:
        example: admin
        type: string
      name:
        example: Administer
        type: string
      type:
        example: user
        type: string
    type: object
  handler.policySubjectsResponse:
    items:
      $ref: '#/definitions/handler.policyIDSubject'
    type: array
...
  handler.resource:
    properties:
      attribute:
        additionalProperties: true
        type: object
      id:
        example: framework
        type: string
      system:
        example: bk_paas
        type: string
      type:
        example: app
        type: string
    required:
    - attribute
    - id
    - system
    - type
    type: object
...
  handler.subject:
    properties:
      id:
        type: string
      type:
        type: string
    required:
    - id
    - type
    type: object
...
  handler.thinPolicyResponse:
    properties:
      expired_at:
        example: 4102444800
        type: integer
      expression:
        additionalProperties: true
        type: object
      id:
        example: 100
        type: integer
      subject:
        $ref: '#/definitions/handler.policyResponseSubject'
        type: object
      version:
        example: "1"
        type: string
    type: object
...
  util.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
info:
  contact: {}
  description: 蓝鲸权限中心后台服务 API 文档
  license: {}
  title: IAM API
  version: "1.0"
paths:
...
  /api/v1/policy/auth:
    post:
      consumes:
      - application/json
      description: 'eval all the policies queried by conditions: system/subject/action
        and resources[required]'
      operationId: api-policy-auth
      parameters:
      - description: the policy request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            $ref: '#/definitions/handler.authResponse'
      security:
      - AppCode: []
      - AppSecret: []
      summary: policy auth/鉴权
      tags:
      - policy
  /api/v1/policy/auth_by_actions:
    post:
      consumes:
      - application/json
      description: batch auth by actions
      operationId: api-policy-batch-auth-by-actions
      parameters:
      - description: the batch auth by actions request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.authByActionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            $ref: '#/definitions/handler.authByActionsResponse'
      security:
      - AppCode: []
      - AppSecret: []
      summary: batch auth by actions/批量鉴权接口
      tags:
      - policy
  /api/v1/policy/auth_by_resources:
    post:
      consumes:
      - application/json
      description: batch auth by resources
      operationId: api-policy-batch-auth-by-resources
      parameters:
      - description: the batch auth by resources request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.authByResourcesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            $ref: '#/definitions/handler.authByResourcesResponse'
      security:
      - AppCode: []
      - AppSecret: []
      summary: batch auth by resources/批量鉴权接口
      tags:
      - policy
...
  /api/v1/systems/{system_id}/policies:
    get:
      consumes:
      - application/json
      description: list policies of a action
      operationId: api-open-system-policies-list
      parameters:
      - description: System ID
        in: path
        name: system_id
        required: true
        type: string
      - in: query
        name: actionID
        required: true
        type: string
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.policyListResponse'
              type: object
      security:
      - AppCode: []
      - AppSecret: []
      summary: policy list
      tags:
      - open
  /api/v1/systems/{system_id}/policies/-/subjects:
    get:
      consumes:
      - application/json
      description: system+action+policy_ids => subjects
      operationId: api-open-system-policies-subjects
      parameters:
      - description: System ID
        in: path
        name: system_id
        required: true
        type: string
      - in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.policySubjectsResponse'
              type: object
      security:
      - AppCode: []
      - AppSecret: []
      summary: query subjects/获取有权限的用户列表
      tags:
      - open
...
securityDefinitions:
  AppCode:
    in: header
    name: X-Bk-App-Code
    type: apiKey
  AppSecret:
    in: header
    name: X-Bk-App-Secret
    type: apiKey
swagger: "2.0"